+++ models/ghosts/Inky.cs
@@ -11,7 +11,7 @@ public class Inky: Ghost
     public static readonly Sprite InkySprite = new Sprite("InkySprite", InkyBitmap, InkyMovingScript);
    
     
-    public Inky(float x, float y, float size, Vector2 velocity, Sprite sprite) : base(x, y, size, velocity, InkySprite)
+    public Inky(int x, int y, int size, Vector2 velocity, Sprite sprite) : base(x, y, size, velocity, InkySprite)
     {
         InkyBitmap.SetCellDetails(InkySprite.Width/8, InkySprite.Height, 8, 1, 8 );
         Strategy = new BlinkyStrategy();
