+++ models/ghosts/Pinky.cs
@@ -11,7 +11,7 @@ public class Pinky: Ghost
     public static readonly Sprite PinkySprite = new Sprite("PinkySprite", PinkyBitmap, PinkyMovingScript);
    
     
-    public Pinky(float x, float y, float size, Vector2 velocity, Sprite sprite) : base(x, y, size, velocity, PinkySprite)
+    public Pinky(int x, int y, int size, Vector2 velocity, Sprite sprite) : base(x, y, size, velocity, PinkySprite)
     {
         PinkyBitmap.SetCellDetails(PinkySprite.Width/8, PinkySprite.Height, 8, 1, 8 );
         Strategy = new BlinkyStrategy();
